+++ lib/adhearsion/call.rb
@@ -30,7 +30,7 @@ module Adhearsion
       end
     end
 
-    attr_reader :end_reason, :commands, :controllers, :variables, :start_time, :end_time
+    attr_reader :end_reason, :controllers, :variables, :start_time, :end_time
 
     delegate :[], :[]=, :to => :variables
     delegate :to, :from, :to => :offer, :allow_nil => true
@@ -58,6 +58,25 @@ module Adhearsion
       offer.target_call_id if offer
     end
     alias :to_s :id
+
+    #
+    # @return [String, nil] The domain on which the call resides
+    #
+    def domain
+      offer.domain if offer
+    end
+
+    #
+    # @return [String, nil] The uri at which the call resides
+    #
+    def uri
+      return nil unless id
+      s = ""
+      s << transport << ":" if transport
+      s << id
+      s << "@" << domain if domain
+      s
+    end
 
     #
     # @return [Array] The set of labels with which this call has been tagged.
@@ -124,6 +143,10 @@ module Adhearsion
     end
     alias << deliver_message
 
+    def commands
+      @commands.clone
+    end
+
     # @private
     def register_initial_handlers
       register_event_handler Punchblock::Event::Offer do |offer|
@@ -171,7 +194,7 @@ module Adhearsion
         clear_from_active_calls
         @end_reason = event.reason
         @end_blocker.broadcast event.reason
-        commands.terminate
+        @commands.terminate
         after(Adhearsion.config.platform.after_hangup_lifetime) { terminate }
         throw :pass
       end
@@ -192,7 +215,7 @@ module Adhearsion
     # Registers a callback for when this call is joined to another call or a mixer
     #
     # @param [Call, String, Hash, nil] target the target to guard on. May be a Call object, a call ID (String, Hash) or a mixer name (Hash)
-    # @option target [String] call_id The call ID to guard on
+    # @option target [String] call_uri The call ID to guard on
     # @option target [String] mixer_name The mixer name to guard on
     #
     def on_joined(target = nil, &block)
@@ -206,7 +229,7 @@ module Adhearsion
     # Registers a callback for when this call is unjoined from another call or a mixer
     #
     # @param [Call, String, Hash, nil] target the target to guard on. May be a Call object, a call ID (String, Hash) or a mixer name (Hash)
-    # @option target [String] call_id The call ID to guard on
+    # @option target [String] call_uri The call ID to guard on
     # @option target [String] mixer_name The mixer name to guard on
     #
     def on_unjoined(target = nil, &block)
@@ -264,7 +287,7 @@ module Adhearsion
     # Joins this call to another call or a mixer
     #
     # @param [Call, String, Hash] target the target to join to. May be a Call object, a call ID (String, Hash) or a mixer name (Hash)
-    # @option target [String] call_id The call ID to join to
+    # @option target [String] call_uri The call ID to join to
     # @option target [String] mixer_name The mixer to join to
     # @param [Hash, Optional] options further options to be joined with
     #
@@ -278,7 +301,7 @@ module Adhearsion
     # Unjoins this call from another call or a mixer
     #
     # @param [Call, String, Hash] target the target to unjoin from. May be a Call object, a call ID (String, Hash) or a mixer name (Hash)
-    # @option target [String] call_id The call ID to unjoin from
+    # @option target [String] call_uri The call ID to unjoin from
     # @option target [String] mixer_name The mixer to unjoin from
     #
     def unjoin(target)
@@ -291,11 +314,11 @@ module Adhearsion
     def join_options_with_target(target)
       case target
       when Call
-        { :call_uri => target.id }
+        { :call_uri => target.uri }
       when String
-        { :call_uri => target }
+        { :call_uri => "#{transport}:#{target}@#{domain}" }
       when Hash
-        abort ArgumentError.new "You cannot specify both a call ID and mixer name" if target.has_key?(:call_id) && target.has_key?(:mixer_name)
+        abort ArgumentError.new "You cannot specify both a call URI and mixer name" if target.has_key?(:call_uri) && target.has_key?(:mixer_name)
         target
       else
         abort ArgumentError.new "Don't know how to join to #{target.inspect}"
@@ -326,10 +349,11 @@ module Adhearsion
 
     # @private
     def write_and_await_response(command, timeout = 60)
-      commands << command
+      @commands << command
       write_command command
 
-      case (response = command.response timeout)
+      response = defer { command.response timeout }
+      case response
       when Punchblock::ProtocolError
         if response.name == :item_not_found
           abort Hangup.new(@end_reason)
@@ -343,6 +367,8 @@ module Adhearsion
       command
     rescue Timeout::Error
       abort CommandTimeout.new(command.to_s)
+    ensure
+      @commands.delete command
     end
 
     # @private
@@ -350,20 +376,19 @@ module Adhearsion
       abort Hangup.new(@end_reason) unless active? || command.is_a?(Punchblock::Command::Hangup)
       merge_headers command.headers if command.respond_to? :headers
       logger.debug "Executing command #{command.inspect}"
-      client.execute_command command, :call_id => id, :async => true
+      client.execute_command command, call_id: id, domain: domain, async: true
     end
 
     # @private
     def logger_id
-      "#{self.class}: #{id}"
+      "#{self.class}: #{id}@#{domain}"
     end
-
     # @private
     def inspect
       attrs = [:offer, :end_reason, :commands, :variables, :controllers, :to, :from].map do |attr|
         "#{attr}=#{send(attr).inspect}"
       end
-      "#<#{self.class}:#{id} #{attrs.join ', '}>"
+      "#<#{self.class}:#{id}@#{domain} #{attrs.join ', '}>"
     end
 
     def execute_controller(controller = nil, completion_callback = nil, &block)
@@ -398,6 +423,10 @@ module Adhearsion
       @client
     end
 
+    def transport
+      offer.transport if offer
+    end
+
     def merge_headers(headers)
       headers.each do |name, value|
         variables[name.to_s.downcase.gsub('-', '_')] = value
@@ -409,7 +438,7 @@ module Adhearsion
     end
 
     # @private
-    class CommandRegistry < ThreadSafeArray
+    class CommandRegistry < Array
       def terminate
         hangup = Hangup.new
         each { |command| command.response = hangup if command.requested? }