+++ lib/adhearsion/call_controller.rb
@@ -54,7 +54,11 @@ module Adhearsion
       end
     end
 
-    attr_reader :call, :metadata
+    # @return [Call] The call object on which the controller is executing
+    attr_reader :call
+
+    # @return [Hash] The controller's metadata provided at invocation
+    attr_reader :metadata
 
     # @private
     attr_reader :block
@@ -170,18 +174,6 @@ module Adhearsion
       @after_call ||= execute_callbacks :after_call
     end
 
-    #
-    # Hangup the call, and execute after_call callbacks
-    #
-    # @param [Hash] headers
-    #
-    def hangup(headers = nil)
-      block_until_resumed
-      call.hangup headers
-      after_call
-      raise Call::Hangup
-    end
-
     # @private
     def write_and_await_response(command)
       block_until_resumed
@@ -210,6 +202,17 @@ module Adhearsion
     end
 
     #
+    # Hangup the call, and execute after_call callbacks
+    #
+    # @param [Hash] headers
+    #
+    def hangup(headers = nil)
+      block_until_resumed
+      call.hangup headers
+      raise Call::Hangup
+    end
+
+    #
     # Reject the call
     #
     # @see Call#reject
@@ -217,6 +220,7 @@ module Adhearsion
     def reject(*args)
       block_until_resumed
       call.reject(*args)
+      raise Call::Hangup
     end
 
     #