+++ lib/adhearsion/punchblock_plugin.rb
@@ -20,8 +20,6 @@ module Adhearsion
       port                Proc.new { PunchblockPlugin.default_port_for_platform platform }, :transform => Proc.new { |v| PunchblockPlugin.validate_number v }, :desc => "Port punchblock needs to connect"
       certs_directory     nil              , :desc => "Directory containing certificates for securing the connection."
       root_domain         nil              , :desc => "The root domain at which to address the server"
-      calls_domain        nil              , :desc => "The domain at which to address calls"
-      mixers_domain       nil              , :desc => "The domain at which to address mixers"
       connection_timeout  60               , :transform => Proc.new { |v| PunchblockPlugin.validate_number v }, :desc => "The amount of time to wait for a connection"
       reconnect_attempts  1.0/0.0          , :transform => Proc.new { |v| PunchblockPlugin.validate_number v }, :desc => "The number of times to (re)attempt connection to the server"
       reconnect_timer     5                , :transform => Proc.new { |v| PunchblockPlugin.validate_number v }, :desc => "Delay between connection attempts"