+++ spec/adhearsion/call_spec.rb
@@ -12,15 +12,19 @@ module Adhearsion
   describe Call do
     let(:mock_client) { double('Client').as_null_object }
 
-    let(:call_id) { rand }
+    let(:call_id) { rand.to_s }
+    let(:domain)  { 'rayo.net' }
     let(:headers) { nil }
     let(:to)      { 'sip:you@there.com' }
     let(:from)    { 'sip:me@here.com' }
+    let(:transport) { 'footransport' }
     let :offer do
-      Punchblock::Event::Offer.new :target_call_id => call_id,
-                                   :to      => to,
-                                   :from    => from,
-                                   :headers => headers
+      Punchblock::Event::Offer.new target_call_id: call_id,
+                                   domain: domain,
+                                   transport: transport,
+                                   to: to,
+                                   from: from,
+                                   headers: headers
     end
 
     subject { Adhearsion::Call.new offer }
@@ -41,13 +45,39 @@ module Adhearsion
     its(:commands) { should be_empty }
 
     its(:id)      { should be == call_id }
+    its(:domain)  { should be == domain }
+    its(:uri)     { should be == "footransport:#{call_id}@#{domain}" }
     its(:to)      { should be == to }
     its(:from)    { should be == from }
+
+    context "when the ID is nil" do
+      let(:call_id) { nil }
+
+      its(:uri) { should be == nil }
+    end
+
+    context "when the domain is nil" do
+      let(:domain) { nil }
+
+      its(:uri) { should be == "footransport:#{call_id}" }
+    end
+
+    context "when the transport is nil" do
+      let(:transport) { nil }
+
+      its(:uri) { should be == "#{call_id}@#{domain}" }
+    end
 
     it "should mark its start time" do
       base_time = Time.local(2008, 9, 1, 12, 0, 0)
       Timecop.freeze base_time
       subject.start_time.should == base_time
+    end
+
+    describe "#commands" do
+      it "should use a duplicating accessor for the command registry" do
+        subject.commands.should_not be subject.commands
+      end
     end
 
     describe "its variables" do
@@ -121,17 +151,32 @@ module Adhearsion
     end
 
     describe "event handlers" do
+      before { pending }
       let(:response) { double 'Response' }
 
       describe "for joined events" do
         context "joined to another call" do
           let :event do
-            Punchblock::Event::Joined.new call_uri: 'foobar'
+            Punchblock::Event::Joined.new call_uri: 'xmpp:foobar@rayo.net'
           end
 
           it "should trigger any on_joined callbacks set for the matching call ID" do
             response.should_receive(:call).once.with(event)
-            subject.on_joined(:call_id => 'foobar') { |event| response.call event }
+            subject.on_joined(:call_uri => 'xmpp:foobar@rayo.net') { |event| response.call event }
+            subject << event
+          end
+
+          it "should trigger any on_joined callbacks set for the matching call ID as a string" do
+            response.should_receive(:call).once.with(event)
+            subject.on_joined('foobar') { |event| response.call event }
+            subject << event
+          end
+
+          it "should trigger any on_joined callbacks set for the matching call" do
+            response.should_receive(:call).once.with(event)
+            call = Call.new
+            call.wrapped_object.stub id: 'foobar', domain: 'rayo.net'
+            subject.on_joined(call) { |event| response.call event }
             subject << event
           end
 
@@ -170,18 +215,32 @@ module Adhearsion
             subject.on_joined(:call_id => 'foobar') { |event| response.call event }
             subject << event
           end
+
+          it "should not trigger any on_joined callbacks set for the matching call ID as a string" do
+            response.should_receive(:call).never
+            subject.on_joined('foobar') { |event| response.call event }
+            subject << event
+          end
+
+          it "should not trigger any on_joined callbacks set for the matching call" do
+            response.should_receive(:call).never
+            call = Call.new
+            call.stub :id => 'foobar'
+            subject.on_joined(call) { |event| response.call event }
+            subject << event
+          end
         end
       end
 
       describe "for unjoined events" do
         context "unjoined from another call" do
           let :event do
-            Punchblock::Event::Unjoined.new call_uri: 'foobar'
+            Punchblock::Event::Unjoined.new call_uri: 'xmpp:foobar@rayo.net'
           end
 
           it "should trigger any on_unjoined callbacks set for the matching call ID" do
             response.should_receive(:call).once.with(event)
-            subject.on_unjoined(:call_id => 'foobar') { |event| response.call event }
+            subject.on_unjoined(:call_uri => 'xmpp:foobar@rayo.net') { |event| response.call event }
             subject << event
           end
 
@@ -194,7 +253,7 @@ module Adhearsion
           it "should trigger any on_unjoined callbacks set for the matching call" do
             response.should_receive(:call).once.with(event)
             call = Call.new
-            call.stub :id => 'foobar'
+            call.wrapped_object.stub id: 'foobar', domain: 'rayo.net'
             subject.on_unjoined(call) { |event| response.call event }
             subject << event
           end
@@ -377,8 +436,11 @@ module Adhearsion
         end
 
         it "should instruct the command registry to terminate" do
-          subject.commands.should_receive(:terminate).once
+          command = Punchblock::Command::Answer.new
+          command.request!
+          subject.future.write_and_await_response command
           subject << end_event
+          command.response(1).should be_a Call::Hangup
         end
 
         it "removes itself from the active calls" do
@@ -480,7 +542,7 @@ module Adhearsion
 
       it "should asynchronously write the command to the Punchblock connection" do
         subject.wrapped_object.should_receive(:client).once.and_return mock_client
-        mock_client.should_receive(:execute_command).once.with(mock_command, :call_id => subject.id, :async => true).and_return true
+        mock_client.should_receive(:execute_command).once.with(mock_command, call_id: call_id, domain: domain, async: true).and_return true
         subject.write_command mock_command
       end
 
@@ -517,9 +579,9 @@ module Adhearsion
         subject.write_and_await_response message
       end
 
-      it "adds the command to the registry" do
+      it "removes the command from the registry after execution" do
         subject.write_and_await_response message
-        subject.commands.should_not be_empty
+        subject.commands.should be_empty
       end
 
       it "blocks until a response is received" do
@@ -532,6 +594,26 @@ module Adhearsion
         starting_time = Time.now
         subject.write_and_await_response slow_command
         (Time.now - starting_time).should >= 0.5
+      end
+
+      context "while waiting for a response" do
+        let(:slow_command) { Punchblock::Command::Dial.new }
+
+        before { slow_command.request! }
+
+        it "does not block the whole actor while waiting for a response" do
+          fut = subject.future.write_and_await_response slow_command
+          subject.id.should == call_id
+          slow_command.response = response
+          fut.value
+        end
+
+        it "adds the command to the registry" do
+          subject.future.write_and_await_response slow_command
+          sleep 0.2
+          subject.commands.should_not be_empty
+          subject.commands.first.should be slow_command
+        end
       end
 
       describe "with a successful response" do
@@ -711,18 +793,20 @@ module Adhearsion
 
         context "with a call" do
           let(:call_id) { rand.to_s }
+          let(:domain)  { 'rayo.net' }
+          let(:uri)     { "footransport:#{call_id}@#{domain}" }
           let(:target)  { described_class.new }
 
-          before { target.stub id: call_id }
+          before { target.wrapped_object.stub uri: uri }
 
           it "should send a join command joining to the provided call ID" do
-            expect_join_with_options :call_id => call_id
+            expect_join_with_options call_uri: uri
             subject.join target
           end
 
           context "and direction/media options" do
             it "should send a join command with the correct options" do
-              expect_join_with_options :call_id => call_id, :media => :bridge, :direction => :recv
+              expect_join_with_options :call_id => uri, :media => :bridge, :direction => :recv
               subject.join target, :media => :bridge, :direction => :recv
             end
           end
@@ -732,30 +816,30 @@ module Adhearsion
           let(:target) { rand.to_s }
 
           it "should send a join command joining to the provided call ID" do
-            expect_join_with_options :call_id => target
+            expect_join_with_options call_uri: "footransport:#{target}@#{subject.domain}"
             subject.join target
           end
 
           context "and direction/media options" do
             it "should send a join command with the correct options" do
-              expect_join_with_options :call_id => target, :media => :bridge, :direction => :recv
+              expect_join_with_options :call_uri => "footransport:#{target}@#{subject.domain}", :media => :bridge, :direction => :recv
               subject.join target, :media => :bridge, :direction => :recv
             end
           end
         end
 
-        context "with a call ID as a hash key" do
+        context "with a call URI as a hash key" do
           let(:call_id) { rand.to_s }
-          let(:target)  { { :call_id => call_id } }
+          let(:target)  { { :call_uri => call_id } }
 
           it "should send a join command joining to the provided call ID" do
-            expect_join_with_options :call_id => call_id
+            expect_join_with_options :call_uri => call_id
             subject.join target
           end
 
           context "and direction/media options" do
             it "should send a join command with the correct options" do
-              expect_join_with_options :call_id => call_id, :media => :bridge, :direction => :recv
+              expect_join_with_options :call_uri => call_id, :media => :bridge, :direction => :recv
               subject.join target.merge({:media => :bridge, :direction => :recv})
             end
           end
@@ -781,10 +865,10 @@ module Adhearsion
         context "with a call ID and a mixer name as hash keys" do
           let(:call_id)     { rand.to_s }
           let(:mixer_name)  { rand.to_s }
-          let(:target)      { { :call_id => call_id, :mixer_name => mixer_name } }
+          let(:target)      { { :call_uri => call_id, :mixer_name => mixer_name } }
 
           it "should raise an ArgumentError" do
-            lambda { subject.join target }.should raise_error ArgumentError, /call ID and mixer name/
+            lambda { subject.join target }.should raise_error ArgumentError, /call URI and mixer name/
           end
         end
       end
@@ -798,12 +882,14 @@ module Adhearsion
 
         context "with a call" do
           let(:call_id) { rand.to_s }
+          let(:domain)  { 'rayo.net' }
+          let(:uri)     { "footransport:#{call_id}@#{domain}" }
           let(:target)  { described_class.new }
 
-          before { target.stub id: call_id }
+          before { target.wrapped_object.stub uri: uri }
 
           it "should send an unjoin command unjoining from the provided call ID" do
-            expect_unjoin_with_options :call_id => call_id
+            expect_unjoin_with_options call_uri: "footransport:#{call_id}@#{domain}"
             subject.unjoin target
           end
         end
@@ -812,17 +898,17 @@ module Adhearsion
           let(:target) { rand.to_s }
 
           it "should send an unjoin command unjoining from the provided call ID" do
-            expect_unjoin_with_options :call_id => target
+            expect_unjoin_with_options call_uri: "footransport:#{target}@#{subject.domain}"
             subject.unjoin target
           end
         end
 
-        context "with a call ID as a hash key" do
+        context "with a call URI as a hash key" do
           let(:call_id) { rand.to_s }
-          let(:target)  { { :call_id => call_id } }
+          let(:target)  { { call_uri: call_id } }
 
           it "should send an unjoin command unjoining from the provided call ID" do
-            expect_unjoin_with_options :call_id => call_id
+            expect_unjoin_with_options call_uri: call_id
             subject.unjoin target
           end
         end
@@ -837,13 +923,13 @@ module Adhearsion
           end
         end
 
-        context "with a call ID and a mixer name as hash keys" do
+        context "with a call URI and a mixer name as hash keys" do
           let(:call_id)     { rand.to_s }
           let(:mixer_name)  { rand.to_s }
-          let(:target)      { { :call_id => call_id, :mixer_name => mixer_name } }
+          let(:target)      { { call_uri: call_id, mixer_name: mixer_name } }
 
           it "should raise an ArgumentError" do
-            lambda { subject.unjoin target }.should raise_error ArgumentError, /call ID and mixer name/
+            lambda { subject.unjoin target }.should raise_error ArgumentError, /call URI and mixer name/
           end
         end
       end